// sivoy-back/src/helpers/serviceType.enum.ts
export enum ServiceType {
  BeachResort = 'Balneario',
  Gastronomy = 'Gastronomía',
  Accommodation = 'Alojamiento',
  Transportation = 'Transporte',
  Entertainment = 'Entretenimiento',
  Culture = 'Cultura',
  Nature = 'Naturaleza',
  AdventureTourism = 'Turismo de Aventura',
  Commerce = 'Comercio',
  Education = 'Educación',
  Sports = 'Deportes',
  Technology = 'Tecnología',
  Professionals = 'Profesionales',
  PlaceOfInterest = 'Lugar de Interés',
  Experiences = 'Experiencias',
  Art = 'Arte',
  Wellness = 'Bienestar',
  Fashion = 'Moda',
  Other = 'Otro',
}